+++ templates/base.html
@@ -18,7 +18,7 @@
         <div class="d-flex">
         <div class="d-flex nav-logo mr-2">
                 <h5>
-                    <a href="{{ url_for('index') }}">
+                    <a href="{{ url_for('index.index') }}">
                         <strong>
                             <i class="fas fa-glass-martini-alt"></i> 
                             <span class="d-none d-md-inline">Drinkit</span>
@@ -39,13 +39,13 @@
     </div>
         <div class="d-flex">
                 <span class="nav-button-3" onclick="toggleSearch()"><i class="fas fa-search"></i></span>
-                <span class="nav-button-1 btn-active d-none d-md-block"><a href="{{ url_for('index', filter='recent') }}"><i class="fas fa-hourglass-start"></i></a></span>
-                <span class="nav-button-2 d-none d-md-block"><a href="{{ url_for('index', filter='popular' ) }}"><i class="fas fa-chart-line"></i></a></span>
+                <span class="nav-button-1 btn-active d-none d-md-block"><a href="{{ url_for('index.index', filter='recent') }}"><i class="fas fa-hourglass-start"></i></a></span>
+                <span class="nav-button-2 d-none d-md-block"><a href="{{ url_for('index.index', filter='popular' ) }}"><i class="fas fa-chart-line"></i></a></span>
             {% if session['username'] %}
                 <span class="nav-button-3 d-none d-md-block">
                     <a href="{{ url_for('view_user_profile', user_id=session['_id']) }}">{{ session['username'] }}</a>
                 </span>
-                <span class="nav-button-3 d-none d-md-block" ><a href="{{ url_for('logout') }}">Log Out</a></span>
+                <span class="nav-button-3 d-none d-md-block" ><a href="{{ url_for('login_logout.logout') }}">Log Out</a></span>
             {% else %}
                 <span class="nav-button-4 d-none d-md-block" data-toggle="modal" data-target="#login">Login</span>
                 <span class="nav-button-4 d-none d-md-block" data-toggle="modal" data-target="#sign-up">Sign-Up</span>
@@ -55,9 +55,9 @@
         <div id="context-menu" class="collapse navbar-collapse pt-3 d-md-none">
             <div class="d-flex flex-column">
                 <strong>SORT:</strong>
-                <span class="nav-button-2 d-md-none btn-active"><a href="{{ url_for('index', filter='recent') }}">
+                <span class="nav-button-2 d-md-none btn-active"><a href="{{ url_for('index.index', filter='recent') }}">
                     <i class="fas fa-hourglass-start"></i>Most Recent</a></span>
-                <span class="nav-button-2 d-md-none"><a href="{{ url_for('index', filter='popular' ) }}"><i
+                <span class="nav-button-2 d-md-none"><a href="{{ url_for('index.index', filter='popular' ) }}"><i
                             class="fas fa-chart-line"></i>Most Popular</a></span>
                 <hr>
                 <strong>USERS:</strong>
@@ -65,9 +65,9 @@
                 <span class="nav-button-2 d-md-none">
                     <a href="{{ url_for('view_user_profile', user_id=session['_id']) }}">{{ session['username'] }}</a>
                 </span>
-                <span class="nav-button-2 d-md-none"><a href="{{ url_for('logout') }}">Log Out</a></span>
+                <span class="nav-button-2 d-md-none"><a href="{{ url_for('login_logout.logout') }}">Log Out</a></span>
                 {% else %}
-                <span class="nav-button-2 d-block d-md-none" data-toggle="modal" data-target="#login">Login</span>
+                <span class="nav-button-2 d-block d-md-none" data-toggle="modal" data-target="login">Login</span>
                 <span class="nav-button-2 d-block d-md-none" data-toggle="modal" data-target="#sign-up">Sign-Up</span>
                 {% endif %}
             </div>
@@ -82,7 +82,7 @@
             <div class="modal-content">
                 <h4>login</h4>
                 
-                <form action="{{ url_for('login') }}" method="POST">
+                <form action="{{ url_for('login_logout.login') }}" method="POST">
                     <div class="form-group">
                         <label for="username">Please Enter Your Username</label>
                         <input id="username" class="modal-input" name="username" type="text" aria-describedby="username"
@@ -115,7 +115,7 @@
                     </div>
                 </div>
                 <div class="row mx-0 px-2">
-                    <form action="{{ url_for('new_user') }}" method="POST" onsubmit="validatePasswords()">
+                    <form action="{{ url_for('login_logout.new_user') }}" method="POST" onsubmit="validatePasswords()">
                         <div class="form-group">
                             <label for="new-username">Please Enter Your Desired Username</label>
                             <input id="new-username" name="newusername" class="modal-input" type="text" placeholder="new username" required>
